requirement.txt
alembic==1.11.1
aniso8601==9.0.1
blinker==1.6.2
click==8.1.6
colorama==0.4.6
Flask==2.3.2
flask-marshmallow==0.15.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.20.1
mysqlclient==2.2.0
packaging==23.1
PyJWT==2.8.0
pytz==2023.3
six==1.16.0
SQLAlchemy==2.0.19
typing_extensions==4.7.1
Werkzeug==2.3.6
